Refactoring and adding GameBase
