+++ TicTacToeClient/nbproject/private/private.xml
@@ -3,14 +3,16 @@
     <editor-bookmarks xmlns="http://www.netbeans.org/ns/editor-bookmarks/2" lastBookmarkId="0"/>
     <open-files xmlns="http://www.netbeans.org/ns/projectui-open-files/2">
         <group>
-<<<<<<< HEAD
-            <file>file:/D:/CSIS/ITI-MAD44/Developing%20Application%20using%20Java%20SE/Project/Java-TicTacToe/TicTacToeClient/src/tictactoeclient/TicTacToeClient.java</file>
-            <file>file:/D:/CSIS/ITI-MAD44/Developing%20Application%20using%20Java%20SE/Project/Java-TicTacToe/TicTacToeClient/src/tictactoe_bord_game/XOBordUI.java</file>
-            <file>file:/D:/CSIS/ITI-MAD44/Developing%20Application%20using%20Java%20SE/Project/Java-TicTacToe/TicTacToeClient/src/tictactoeclient_computer_game/XOBordComputer.java</file>
-=======
-            <file>file:/C:/Users/Hp/Desktop/Java%20Project/TicTacToeClient/src/tictactoeclient/TicTacToeClient.java</file>
-            <file>file:/C:/Users/Hp/Desktop/Java%20Project/TicTacToeClient/src/tictactoe_bord_game/XOBordUI.java</file>
->>>>>>> 6cc77e94e82ae8ff1f7ca2566d2098cf78a02572
+            <file>file:/C:/Users/Hp/Desktop/Java%20Project/TicTacToeClient/src/home/EasyHardBase.java</file>
+            <file>file:/C:/Users/Hp/Desktop/Java%20Project/TicTacToeClient/src/boardGamePkg/XOBoardGUI.fxml</file>
+            <file>file:/C:/Users/Hp/Desktop/Java%20Project/TicTacToeClient/src/home/OnlineOfflineScreen.java</file>
+            <file>file:/C:/Users/Hp/Desktop/Java%20Project/TicTacToeClient/src/home/HomeGUIController.java</file>
+            <file>file:/C:/Users/Hp/Desktop/Java%20Project/TicTacToeClient/src/home/TestingGui.java</file>
+            <file>file:/C:/Users/Hp/Desktop/Java%20Project/TicTacToeClient/src/home/AiBase.java</file>
+            <file>file:/C:/Users/Hp/Desktop/Java%20Project/TicTacToeClient/src/home/HomeGUI.fxml</file>
+            <file>file:/C:/Users/Hp/Desktop/Java%20Project/TicTacToeClient/src/boardGamePkg/XOBoardGUIController.java</file>
+            <file>file:/C:/Users/Hp/Desktop/Java%20Project/TicTacToeClient/src/boardGamePkg/LocalMultiMode.java</file>
+            <file>file:/C:/Users/Hp/Desktop/Java%20Project/TicTacToeClient/src/boardGamePkg/LocalSingleMode.java</file>
         </group>
     </open-files>
 </project-private>
