--- TicTacToeClient/src/AiHuman/FXMLDocument.fxml
@@ -1,66 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-
-<?import javafx.geometry.Insets?>
-<?import javafx.scene.control.Button?>
-<?import javafx.scene.control.Label?>
-<?import javafx.scene.image.ImageView?>
-<?import javafx.scene.layout.ColumnConstraints?>
-<?import javafx.scene.layout.GridPane?>
-<?import javafx.scene.layout.RowConstraints?>
-<?import javafx.scene.text.Font?>
-
-<GridPane maxHeight="-Infinity" maxWidth="-Infinity" minHeight="-Infinity" minWidth="-Infinity" prefHeight="400.0" prefWidth="600.0" xmlns="http://javafx.com/javafx/8.0.171" xmlns:fx="http://javafx.com/fxml/1">
-  <columnConstraints>
-      <ColumnConstraints hgrow="SOMETIMES" maxWidth="422.0" minWidth="10.0" prefWidth="29.0" />
-    <ColumnConstraints hgrow="SOMETIMES" maxWidth="422.0" minWidth="10.0" prefWidth="221.0" />
-    <ColumnConstraints hgrow="SOMETIMES" maxWidth="313.0" minWidth="10.0" prefWidth="98.0" />
-      <ColumnConstraints hgrow="SOMETIMES" maxWidth="313.0" minWidth="10.0" prefWidth="252.0" />
-  </columnConstraints>
-  <rowConstraints>
-    <RowConstraints maxHeight="104.0" minHeight="10.0" prefHeight="88.0" vgrow="SOMETIMES" />
-    <RowConstraints maxHeight="104.0" minHeight="10.0" prefHeight="88.0" vgrow="SOMETIMES" />
-    <RowConstraints maxHeight="104.0" minHeight="10.0" prefHeight="88.0" vgrow="SOMETIMES" />
-      <RowConstraints maxHeight="104.0" minHeight="10.0" prefHeight="88.0" vgrow="SOMETIMES" />
-      <RowConstraints maxHeight="104.0" minHeight="10.0" prefHeight="88.0" vgrow="SOMETIMES" />
-      <RowConstraints maxHeight="104.0" minHeight="10.0" prefHeight="88.0" vgrow="SOMETIMES" />
-      <RowConstraints maxHeight="129.0" minHeight="0.0" prefHeight="40.0" vgrow="SOMETIMES" />
-      <RowConstraints maxHeight="246.0" minHeight="10.0" prefHeight="53.0" vgrow="SOMETIMES" />
-      <RowConstraints maxHeight="246.0" minHeight="10.0" prefHeight="219.0" vgrow="SOMETIMES" />
-  </rowConstraints>
-   <children>
-      <Label prefHeight="93.0" prefWidth="112.0" text="Tac" GridPane.columnIndex="2" GridPane.rowIndex="3">
-         <font>
-            <Font size="64.0" />
-         </font>
-      </Label>
-      <Label layoutX="50.0" layoutY="2.0" prefHeight="93.0" prefWidth="224.0" text="        Tic" textAlignment="CENTER" textOverrun="CENTER_ELLIPSIS" GridPane.columnIndex="1" GridPane.columnSpan="2" GridPane.rowIndex="2">
-         <font>
-            <Font size="64.0" />
-         </font>
-      </Label>
-      <ImageView fitHeight="150.0" fitWidth="200.0" pickOnBounds="true" preserveRatio="true" />
-      <Button alignment="TOP_LEFT" contentDisplay="TOP" mnemonicParsing="false" text="&lt;-" textOverrun="LEADING_ELLIPSIS">
-         <font>
-            <Font size="24.0" />
-         </font>
-         <padding>
-            <Insets bottom="1.0" left="1.0" right="1.0" top="1.0" />
-         </padding>
-      </Button>
-      <Label text="Teo" GridPane.columnIndex="3" GridPane.rowIndex="4">
-         <font>
-            <Font size="64.0" />
-         </font>
-      </Label>
-      <Button mnemonicParsing="false" prefHeight="52.0" prefWidth="215.0" text="Human" GridPane.columnIndex="1" GridPane.rowIndex="8">
-         <font>
-            <Font name="System Bold Italic" size="36.0" />
-         </font>
-      </Button>
-      <Button mnemonicParsing="false" prefHeight="80.0" prefWidth="212.0" text="AI" GridPane.columnIndex="3" GridPane.rowIndex="8">
-         <font>
-            <Font name="System Bold Italic" size="36.0" />
-         </font>
-      </Button>
-   </children>
-</GridPane>
