+++ TicTacToeClient/src/boardGamePkg/XOBoardGUI.fxml
@@ -2,6 +2,7 @@
 
 <?import javafx.scene.control.Button?>
 <?import javafx.scene.control.Label?>
+<?import javafx.scene.control.ToggleButton?>
 <?import javafx.scene.layout.Pane?>
 <?import javafx.scene.shape.Line?>
 <?import javafx.scene.text.Font?>
@@ -32,5 +33,8 @@
       <Button layoutX="324.0" layoutY="92.0" mnemonicParsing="false" prefHeight="70.0" prefWidth="95.0" />
       <Button layoutX="324.0" layoutY="173.0" mnemonicParsing="false" prefHeight="70.0" prefWidth="95.0" />
       <Button layoutX="324.0" layoutY="250.0" mnemonicParsing="false" prefHeight="70.0" prefWidth="95.0" />
+      <Label fx:id="scoreP1" layoutX="175.0" layoutY="48.0" />
+      <Label fx:id="scoreP2" layoutX="344.0" layoutY="48.0" />
+      <ToggleButton fx:id="recordBtn" layoutX="14.0" layoutY="361.0" mnemonicParsing="false" text="Record" />
    </children>
 </Pane>
