+++ TicTacToeClient/src/boardGamePkg/XOBoardGUIController.java
@@ -1,11 +1,9 @@
-package SignIn;
-
 /*
  * To change this license header, choose License Headers in Project Properties.
  * To change this template file, choose Tools | Templates
  * and open the template in the editor.
  */
-
+package boardGamePkg;
 
 import java.net.URL;
 import java.util.ResourceBundle;
@@ -16,7 +14,7 @@ import javafx.fxml.Initializable;
  *
  * @author Hp
  */
-public class SignInController implements Initializable {
+public class XOBoardGUIController implements Initializable {
 
     /**
      * Initializes the controller class.
