+++ TicTacToeClient/src/home/HomeBaseGUI.fxml
@@ -1,30 +1,30 @@
 <?xml version="1.0" encoding="UTF-8"?>
 
+<?import javafx.geometry.Insets?>
 <?import javafx.scene.control.Button?>
 <?import javafx.scene.control.Label?>
-<?import javafx.scene.image.ImageView?>
 <?import javafx.scene.layout.ColumnConstraints?>
 <?import javafx.scene.layout.GridPane?>
 <?import javafx.scene.layout.RowConstraints?>
 <?import javafx.scene.text.Font?>
 
 <GridPane maxHeight="-Infinity" maxWidth="-Infinity" minHeight="-Infinity" minWidth="-Infinity" prefHeight="400.0" prefWidth="600.0" xmlns="http://javafx.com/javafx/8.0.171" xmlns:fx="http://javafx.com/fxml/1">
-  <columnConstraints>
+   <columnConstraints>
       <ColumnConstraints hgrow="SOMETIMES" maxWidth="422.0" minWidth="10.0" prefWidth="58.0" />
-    <ColumnConstraints hgrow="SOMETIMES" maxWidth="422.0" minWidth="10.0" prefWidth="189.0" />
-    <ColumnConstraints hgrow="SOMETIMES" maxWidth="313.0" minWidth="10.0" prefWidth="97.0" />
+      <ColumnConstraints hgrow="SOMETIMES" maxWidth="422.0" minWidth="10.0" prefWidth="189.0" />
+      <ColumnConstraints hgrow="SOMETIMES" maxWidth="313.0" minWidth="10.0" prefWidth="97.0" />
       <ColumnConstraints hgrow="SOMETIMES" maxWidth="313.0" minWidth="10.0" prefWidth="256.0" />
-  </columnConstraints>
-  <rowConstraints>
-    <RowConstraints maxHeight="104.0" minHeight="10.0" prefHeight="88.0" vgrow="SOMETIMES" />
-    <RowConstraints maxHeight="104.0" minHeight="10.0" prefHeight="88.0" vgrow="SOMETIMES" />
-    <RowConstraints maxHeight="104.0" minHeight="10.0" prefHeight="88.0" vgrow="SOMETIMES" />
+   </columnConstraints>
+   <rowConstraints>
+      <RowConstraints maxHeight="104.0" minHeight="10.0" prefHeight="88.0" vgrow="SOMETIMES" />
+      <RowConstraints maxHeight="104.0" minHeight="10.0" prefHeight="88.0" vgrow="SOMETIMES" />
+      <RowConstraints maxHeight="104.0" minHeight="10.0" prefHeight="88.0" vgrow="SOMETIMES" />
       <RowConstraints maxHeight="104.0" minHeight="10.0" prefHeight="88.0" vgrow="SOMETIMES" />
       <RowConstraints maxHeight="104.0" minHeight="10.0" prefHeight="88.0" vgrow="SOMETIMES" />
       <RowConstraints maxHeight="129.0" minHeight="0.0" prefHeight="40.0" vgrow="SOMETIMES" />
       <RowConstraints maxHeight="246.0" minHeight="10.0" prefHeight="53.0" vgrow="SOMETIMES" />
       <RowConstraints maxHeight="246.0" minHeight="10.0" prefHeight="219.0" vgrow="SOMETIMES" />
-  </rowConstraints>
+   </rowConstraints>
    <children>
       <Label prefHeight="93.0" prefWidth="112.0" text="Tac" GridPane.columnIndex="2" GridPane.rowIndex="2">
          <font>
@@ -36,7 +36,6 @@
             <Font size="64.0" />
          </font>
       </Label>
-      <ImageView fitHeight="150.0" fitWidth="200.0" pickOnBounds="true" preserveRatio="true" />
       <Label text="Teo" GridPane.columnIndex="3" GridPane.rowIndex="3">
          <font>
             <Font size="64.0" />
@@ -51,6 +50,11 @@
          <font>
             <Font name="System Bold Italic" size="36.0" />
          </font>
+      </Button>
+      <Button fx:id="backBtn" mnemonicParsing="false" text="Back">
+         <GridPane.margin>
+            <Insets left="13.5" top="12.0" />
+         </GridPane.margin>
       </Button>
    </children>
 </GridPane>
