+++ TicTacToeClient/src/signInPkg/SignInController.java
@@ -1,9 +1,11 @@
+package signInPkg;
+
 /*
  * To change this license header, choose License Headers in Project Properties.
  * To change this template file, choose Tools | Templates
  * and open the template in the editor.
  */
-package SignUp;
+
 
 import java.net.URL;
 import java.util.ResourceBundle;
@@ -14,7 +16,7 @@ import javafx.fxml.Initializable;
  *
  * @author Hp
  */
-public class SignUPUIController implements Initializable {
+public class SignInController implements Initializable {
 
     /**
      * Initializes the controller class.
